requirements.txt
fastapi
uvicorn
pydantic
httpx
python-dotenv
deepseek-ai
asyncio